Move code to separate files to improve readability
